+++ cicerone/app/src/main/java/com/teampower/cicerone/GeofenceTriggeredActivity.kt
@@ -7,7 +7,7 @@ import android.text.Html
 import android.text.method.LinkMovementMethod
 import android.util.Log
 import androidx.lifecycle.ViewModelProvider
-import com.teampower.cicerone.database.category_table.CategoryViewModel
+import com.teampower.cicerone.viewmodels.CategoryViewModel
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import com.google.gson.internal.LinkedTreeMap
