+++ cicerone/app/src/main/java/com/teampower/cicerone/database/CategoryDao.kt
@@ -1,4 +1,4 @@
-package com.teampower.cicerone.database.category_table
+package com.teampower.cicerone.database
 
 import androidx.lifecycle.LiveData
 import androidx.room.*
