+++ cicerone/app/src/main/java/com/teampower/cicerone/viewmodels/POIHistoryViewModel.kt
@@ -1,11 +1,12 @@
-package com.teampower.cicerone.database.category_table
+package com.teampower.cicerone.viewmodels
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.viewModelScope
-import com.teampower.cicerone.database.CategoryData
 import com.teampower.cicerone.database.CiceroneAppDatabase
+import com.teampower.cicerone.database.POIHistoryData
+import com.teampower.cicerone.database.repositories.POIHistoryRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
@@ -15,26 +16,26 @@ import kotlinx.coroutines.launch
 * while your ViewModel can take care of holding and processing all the data needed for the UI.
 * */
 
-class CategoryViewModel(application: Application) : AndroidViewModel(application) {
+class POIHistoryViewModel(application: Application) : AndroidViewModel(application) {
 
-    private val repository: CategoryRepository
+    private val historyRepository: POIHistoryRepository
 
     // Using LiveData and caching what getAlphabetizedWords returns has several benefits:
     // - We can put an observer on the data (instead of polling for changes) and only update the
     //   the UI when the data actually changes.
     // - Repository is completely separated from the UI through the ViewModel.
-    val allCat: LiveData<List<CategoryData>>
+    val allPOI: LiveData<List<POIHistoryData>>
 
     init {
-        val catDao = CiceroneAppDatabase.getDatabase(
+        val poisDao = CiceroneAppDatabase.getDatabase(
             application,
             viewModelScope
-        ).categoryDao()
-        repository =
-            CategoryRepository(
-                catDao
+        ).poiHistoryDao()
+        historyRepository =
+            POIHistoryRepository(
+                poisDao
             )
-        allCat = repository.allCatScores
+        allPOI = historyRepository.allPOI
     }
 
     /**
@@ -42,15 +43,7 @@ class CategoryViewModel(application: Application) : AndroidViewModel(application
      * to block the main thread, so we're launching a new coroutine and calling the repository's
      * insert, which is a suspend function.
      */
-    fun insert(cat: CategoryData) = viewModelScope.launch(Dispatchers.IO) {
-        repository.insert(cat)
-    }
-
-    fun updateCategoryPoints(catName: String, points: Double) = viewModelScope.launch(Dispatchers.IO) {
-        repository.updateCategoryPoints(catName, points)
-    }
-
-    fun getCategoryPoints(catName: String) = viewModelScope.launch(Dispatchers.IO) {
-        repository.getCategoryPoints(catName)
+    fun insert(poi: POIHistoryData) = viewModelScope.launch(Dispatchers.IO) {
+        historyRepository.insert(poi)
     }
 }
